# docs/api-docs.yml
openapi: 3.0.1
info:
  title: Bestellsystem API
  version: v0
servers:
  - url: http://localhost:3000
    description: Generated server url
paths:
...
  /orders:
    get:
      tags:
        - Orders
      operationId: getAllOrders
      parameters:
        - in: query
          name: owner_id
          schema:
            type: string
          required: false
          description: Filter orders by the owner's ID.
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Page number of the results to retrieve.
        - in: query
          name: size
          schema:
            type: integer
          required: false
          description: Number of orders to return per page.
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseDTO'
    post:
      tags:
        - Orders
      operationId: createOrder
      description: >
        Erstellt eine neue Bestellung und setzt den `status` auf "In Bearbeitung".
        `booking_year` & `primary_cost_center_id` sind verpflichtend, um eine neue Bestellung anlegen zu können.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderResponseDTO'
  /orders/{order_id}/items:
    get:
      tags:
        - Orders
      operationId: getOrderItems
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
          description: ID of the order to retrieve items for.
      responses:
        '200':
          description: Successfully retrieved items for the order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponseDTO'
        '404':
          description: Bestellung nicht gefunden.
  /orders/{order_id}/quotations:
    get:
      tags:
        - Orders
      operationId: getOrderQuotations
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
          description: ID of the order to retrieve quotations for.
      responses:
        '200':
          description: Successfully retrieved quotations for the order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuotationResponseDTO'
        '404':
          description: Bestellung nicht gefunden.
  /orders/{order_id}/export:
    get:
      tags:
        - Orders
      operationId: exportOrderToFormula
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
          description: ID of the order to export to PDF.
      responses:
        '200':
          description: Successfully retrieved PDF formula for the order.
          content:
            application/pdf: { }
        '404':
          description: Bestellung nicht gefunden.
...
components:
  schemas:
...
    OrderRequestDTO:
      type: object
      required:
        - booking_year
        - primary_cost_center_id
      properties:
        primary_cost_center_id:
          type: string
        booking_year:
          type: string
          minLength: 2
          maxLength: 2
        legacy_alias:
          type: string
        owner_id:
          type: string
        content_description:
          type: string
        currency_short:
          type: string
        comment:
          type: string
        comment_for_supplier:
          type: string
        quote_number:
          type: string
        quote_sign:
          type: string
        quote_date:
          type: string
          format: date
        quote_price:
          type: number
        delivery_person_id:
          type: integer
          format: int64
        invoice_person_id:
          type: integer
          format: int64
        queries_person_id:
          type: integer
          format: int64
        customer_id:
          type: string
        supplier_id:
          type: integer
          format: int32
        secondary_cost_center_id:
          type: string
        fixed_discount:
          type: number
        percentage_discount:
          type: number
        cashback_percentage:
          type: number
        cashback_days:
          type: integer
          format: int32
        last_updated_time:
          type: string
          format: date-time
        flag_decision_cheapest_offer:
          type: boolean
        flag_decision_sole_supplier:
          type: boolean
        flag_decision_contract_partner:
          type: boolean
        flag_decision_other_reasons:
          type: boolean
        decision_other_reasons_description:
          type: string
        flag_edv_permission:
          type: boolean
        flag_furniture_permission:
          type: boolean
        flag_furniture_room:
          type: boolean
        flag_investment_room:
          type: boolean
        flag_investment_structural_measures:
          type: boolean
        flag_media_permission:
          type: boolean
        dfg_key:
          type: string
...
    OrderResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        primary_cost_center_id:
          type: string
        booking_year:
          type: string
        auto_index:
          type: integer
          format: int32
        created_date:
          type: string
          format: date-time
        legacy_alias:
          type: string
        owner_id:
          type: string
        content_description:
          type: string
        status:
          type: string
          enum:
            - ABR
            - ABS
            - ARC
            - DEL
            - INB
        currency:
          $ref: '#/components/schemas/CurrencyResponseDTO'
        comment:
          type: string
        comment_for_supplier:
          type: string
        quote_number:
          type: string
        quote_sign:
          type: string
        quote_date:
          type: string
          format: date
        quote_price:
          type: number
        delivery_person_id:
          type: integer
          format: int64
        invoice_person_id:
          type: integer
          format: int64
        queries_person_id:
          type: integer
          format: int64
        customer_id:
          type: string
        supplier_id:
          type: integer
          format: int32
        secondary_cost_center_id:
          type: string
        fixed_discount:
          type: number
        percentage_discount:
          type: number
        cash_discount:
          type: number
        cashback_days:
          type: integer
          format: int32
        last_updated_time:
          type: string
          format: date-time
        flag_decision_cheapest_offer:
          type: boolean
        flag_decision_sole_supplier:
          type: boolean
        flag_decision_contract_partner:
          type: boolean
        flag_decision_other_reasons:
          type: boolean
        decision_other_reasons_description:
          type: string
        flag_edv_permission:
          type: boolean
        flag_furniture_permission:
          type: boolean
        flag_furniture_room:
          type: boolean
        flag_investment_room:
          type: boolean
        flag_investment_structural_measures:
          type: boolean
        flag_media_permission:
          type: boolean
        dfg_key:
          type: string
...
    VatResponseDTO:
      type: object
      properties:
        value:
          type: number
        description:
          type: string
...
    ItemResponseDTO:
      type: object
      properties:
        order_id:
          type: integer
          format: int64
        item_id:
          type: integer
          format: int32
        name:
          type: string
        price_per_unit:
          type: number
        quantity:
          type: integer
          format: int64
        quantity_unit:
          type: string
        article_id:
          type: string
        comment:
          type: string
        Vat's:
          $ref: '#/components/schemas/VatResponseDTO'
        preferred_list:
          type: string
          enum:
            - RZ
            - TA
        preferred_list_number:
          type: string
        vat_type:
          type: string
          enum:
            - netto
            - brutto
    CurrencyResponseDTO:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
    QuotationResponseDTO:
      type: object
      properties:
        order_id:
          type: integer
          format: int64
        index:
          type: integer
          format: int32
        quote_date:
          type: string
          format: date
        price:
          type: number
        supplier_id:
          type: integer
          format: int64
...